Split state to allow easy extensibility
