+++ src/Character/actions.js
@@ -1,0 +1,7 @@
+export const actions = {
+  SORT_ASC: "SORT_ASC",
+  ALL_CHARACTERS: "ALL_CHARACTERS",
+  CHARACTERS: "CHARACTERS",
+  CHARACTERS_LOADING: "CHARACTERS_LOADING",
+  CHARACTERS_ERROR: "CHARACTERS_ERROR"
+};
