+++ src/Character/reducer.js
@@ -1,6 +1,6 @@
-import { actions } from "./store";
+import { actions } from "./actions";
 
-export const appReducer = (state, action) => {
+export const characterReducer = (state, action) => {
   switch (action.type) {
     case actions.ALL_CHARACTERS:
       return { ...state, allCharacters: action.payload, loading: false };
@@ -8,16 +8,8 @@ export const appReducer = (state, action) => {
       return { ...state, characters: action.payload, loading: false };
     case actions.SORT_ASC:
       return { ...state, sortAscending: !state.sortAscending };
-    case actions.MOVIE:
-      return { ...state, movies: action.payload, loading: false };
-    case actions.SELECTED_MOVIE:
-      return { ...state, selectedMovie: action.payload, loading: false };
-    case actions.MOVIE_LOADING:
-      return { ...state, movieLoading: action.payload };
     case actions.CHARACTERS_LOADING:
       return { ...state, charactersLoading: action.payload };
-    case actions.MOVIE_ERROR:
-      return { ...state, movieError: action.payload };
     case actions.CHARACTERS_ERROR:
       return { ...state, charactersError: action.payload };
     default: {
