+++ src/Movies/actions.js
@@ -1,0 +1,6 @@
+export const actions = {
+  MOVIE_LOADING: "MOVIE_LOADING",
+  MOVIE_ERROR: "MOVIE_ERROR",
+  MOVIE: "MOVIE",
+  SELECTED_MOVIE: "SELECTED_MOVIE"
+};
